adjust deploy and build parameters

+++ lib/Rex/Apache/Deploy/Package/deb.pm
@@ -6,12 +6,18 @@
 
 package Rex::Apache::Deploy::Package::deb;
 
+use strict;
+use warnings;
+
 use Rex::Apache::Deploy::Package::Base;
 use base qw(Rex::Apache::Deploy::Package::Base);
 
+use Rex::Commands;
 use Rex::Commands::Fs;
 use Rex::Commands::Run;
 use Rex::Commands::Upload;
+use Rex::Apache::Build;
+use Data::Dumper;
 
 sub new {
    my $that = shift;
@@ -20,19 +26,32 @@ sub new {
 
    bless($self, $proto);
 
+   if($self->{arch} eq "x86_64") {
+      $self->{arch} = "amd64";
+   }
+
    return $self;
 }
 
 sub deploy {
-   my ($package_name) = @_;
+   my ($self, $package_name, %option) = @_;
 
-   if(! $package_name) {
-      $package_name = $self->name . "_" . $self->version . "_" . $self->arch . ".deb";
-   }
+   $package_name = $self->name . "_" . $self->version . "_" . $self->arch . ".deb";
+
+   LOCAL {
+      if(! -f $package_name) {
+         build($self->name, %option);
+      }
+   };
 
    upload $package_name, "/tmp";
 
    run "dpkg -i /tmp/$package_name";
+   if($? != 0) {
+      die("Error installing $package_name");
+   }
+
+   Rex::Logger::info("Package $package_name installed.");
 
    unlink "/tmp/$package_name";
 }
