+++ lib/Rex/Apache/Deploy/Package/rpm.pm
@@ -9,9 +9,14 @@ package Rex::Apache::Deploy::Package::rpm;
 use Rex::Apache::Deploy::Package::Base;
 use base qw(Rex::Apache::Deploy::Package::Base);
 
+use strict;
+use warnings;
+
+use Rex::Commands;
 use Rex::Commands::Fs;
 use Rex::Commands::Run;
 use Rex::Commands::Upload;
+use Rex::Apache::Build;
 
 sub new {
    my $that = shift;
@@ -20,19 +25,30 @@ sub new {
 
    bless($self, $proto);
 
+   $self->{release}    ||= 1;
+
    return $self;
 }
 
 sub deploy {
-   my ($package_name) = @_;
+   my ($self, $package_name, %option) = @_;
 
-   if(! $package_name) {
-      $package_name = $self->name . "-" . $self->version . "-". $self->release . "." . $self->arch . ".rpm";
-   }
+   $package_name = $self->name . "-" . $self->version . "-". $self->release . "." . $self->arch . ".rpm";
+
+   LOCAL {
+      if(! -f $package_name) {
+         build($self->name, %option);
+      }
+   };
 
    upload $package_name, "/tmp";
 
    run "rpm -U /tmp/$package_name";
+   if($? != 0) {
+      die("Error installing $package_name");
+   }
+
+   Rex::Logger::info("Package $package_name installed.");
 
    unlink "/tmp/$package_name";
 }
